refactor: remove Footer/v2 subfolder
