+++ operate/client/src/App/Processes/MigrationView/Footer/buildMigrationBatchOperationFilter.ts
