+++ operate/client/src/App/Processes/MigrationView/Footer/index.test.tsx
